I believe no more polishing
I hate that you can see the text before it changes to the text I want it to be but since I have no solution and it bothers quite literally no one except me, I'm not gonna bother at the moment.

Also, since I rely on the ContentRedirectionModule SDCafiine also relies on, I can't do a feature I wanted to do that is just not use subfolders at all for files. Might fork ConentRedirectionModule for the plugin or just leave it and hope no one cares :)